requirements.txt
aiohttp
aiofiles
google-auth
dotenv
pygame
python-dotenv
google-genai
pyaudio
opencv-python
sounddevice